Updated names for IENU & SQL Services
