+++ Services/IEnumerable/TrainerDB_IENU.cs
@@ -1,7 +1,7 @@
 ﻿using Gadevang_Tennis_Klub.Interfaces.Services;
 namespace Gadevang_Tennis_Klub.Services.IENumerable
 {
-    public class TrainerDB : ITrainerDB
+    public class TrainerDB_IENU : ITrainerDB
     {
     }
 }
