+++ Services/SQL/Booking/CourtBookingDB_SQL.cs
@@ -3,7 +3,7 @@ using Gadevang_Tennis_Klub.Interfaces.Services;
 
 namespace Gadevang_Tennis_Klub.Services.SQL.Booking
 {
-    public class CourtBookingDB : ICourtBookingDB
+    public class CourtBookingDB_SQL : ICourtBookingDB
     {
         public bool CreateCourtBookingAsync(ICourtBooking courtBooking)
         {
