+++ Services/SQL/TrainerDB_SQL.cs
@@ -1,7 +1,7 @@
 ﻿using Gadevang_Tennis_Klub.Interfaces.Services;
 namespace Gadevang_Tennis_Klub.Services.SQL
 {
-    public class TrainerDB : ITrainerDB
+    public class TrainerDB_SQL : ITrainerDB
     {
     }
 }
